requirements-dev.txt
pytest
pytest-flask
pytest-asyncio
coverage
flake8
mypy
httpx